demo/requirements.txt
streamlit==1.18.1
openai==0.26.5
python-dotenv==0.21.0
langchain